Middleware proved ineffective as multiple actions get dispatched per search and so the middleware needs to be applied to each one. Moved functionality to 'submitSearch' action
